Publishing version 2.1.18.10

+++ org.radixware/kernel/explorer/src/org/radixware/kernel/explorer/editors/sqmleditor/sqmltags/SqmlTag_EndIf.java
@@ -26,7 +26,7 @@ public class SqmlTag_EndIf  extends SqmlTag {
         super(environment,source);
     }
     
-    public SqmlTag_EndIf(final IClientEnvironment environment, final int pos){
+    public SqmlTag_EndIf(final IClientEnvironment environment, final long pos){
         super(environment,pos);
         setDisplayedInfo(null,"#END IF");
     }
